requirements.txt
Flask==2.2.3
PyJWT==2.8.0
requests==2.28.2
